+++ lib/BleFingerprint/rssi.h
@@ -1,19 +1,19 @@
 #ifndef _RSSI_
 #define _RSSI_
 
-#define CLOSE_RSSI (-40)
-#define LEFT_RSSI (-50)
+#define CLOSE_RSSI int8_t(-40)
+#define LEFT_RSSI int8_t(-50)
 
-#define DEFAULT_TX (-6)
+#define DEFAULT_TX int8_t(-6)
 
-#define APPLE_TX 0
-#define RM_ASST_TX 0
-#define TILE_TX (-4)
-#define EXPOSURE_TX (-12)
-#define ITAG_TX (-10)
+#define APPLE_TX int8_t(0)
+#define RM_ASST_TX int8_t(0)
+#define TILE_TX int8_t(-4)
+#define EXPOSURE_TX int8_t(-12)
+#define ITAG_TX int8_t(-10)
 
-#define NUT_TX (-12)
-#define FLORA_TX (-10)
+#define NUT_TX int8_t(-12)
+#define FLORA_TX int8_t(-10)
 
-#define EDDYSTONE_ADD_1M (-41)
+#define EDDYSTONE_ADD_1M int8_t(-41)
 #endif
